[MOVE] <shared module> set file structure
