+++ src/shared/schedule/index.ts
@@ -1,5 +1,5 @@
 import { MealJobScheduler } from "../../event/meal/meal.scheduler";
-import { NoticeJobScheduler } from "../../notice/notice.scheduler";
+import { NoticeJobScheduler } from "../../notice/schedule/notice.scheduler";
 import { JobScheduler } from "./schedule.type";
 
 const setGlobalSchedule = () => {
